feature: AP-9 created navigation hook for more scalabe hook as i had to change paths everywhere, added feature to disable navigation when on it
