+++ src/components/organisms/NavigationMenu/NavigationMenu.tsx
@@ -1,19 +1,13 @@
 import type { ReactNode } from "react";
 import { APDrawer, APNavBar, useAPDrawer } from "../../molecules";
-import { getNavigationPanelItems } from "./constants/navigationPanelItems";
-import { useDispatch } from "react-redux";
-import { useNavigate } from "react-router-dom";
-import { logout } from "@/services/auth";
 import { createUserMenuItems } from "./helpers";
+import { useNavigationMenu } from "./hooks/useNavigationMenu";
 
 interface INavigationMenu {
   children?: ReactNode;
 }
 
 export const NavigationMenu = ({ children }: INavigationMenu) => {
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-
   const {
     mobileOpen,
     handleDrawerClose,
@@ -21,26 +15,10 @@ export const NavigationMenu = ({ children }: INavigationMenu) => {
     handleDrawerToggle,
   } = useAPDrawer();
 
-  const handleLogout = () => {
-    dispatch(logout());
-    navigate("/login");
-  };
-
-  const handleDashboardNavigation = () => {
-    navigate("/dashboard");
-    handleDrawerClose();
-  };
-
-  const handleTableNavigation = () => {
-    navigate("/table");
-    handleDrawerClose();
-  };
+  const { handleLogout, navigationPanelItems } =
+    useNavigationMenu(handleDrawerClose);
 
   const userMenuItems = createUserMenuItems(handleLogout);
-  const navigationPanelItems = getNavigationPanelItems(
-    handleDashboardNavigation,
-    handleTableNavigation
-  );
 
   return (
     <>
