+++ src/components/organisms/NavigationMenu/constants/navigationPath.ts
@@ -1,0 +1,5 @@
+export enum NAVIGATION_PATH {
+  login = "/login",
+  dashboard = "/dashboard",
+  users = "/users",
+}
